refactor ledger api connection to handle contract api
