+++ packages/fetchai/connections/ledger_api/connection.yaml
@@ -6,14 +6,16 @@ license: Apache-2.0
 aea_version: '>=0.4.0, <0.5.0'
 fingerprint:
   __init__.py: QmZvYZ5ECcWwqiNGh8qNTg735wu51HqaLxTSifUxkQ4KGj
-  connection.py: QmaXZt2TZWLRUQJjN7F8pVW5jzXaoVmc6J2F1REEczH5qq
+  base.py: QmYvBYACFLwyjDcZjWfUjzHJypNTJ6duMo51tCGDdAzgJU
+  connection.py: QmUEXAherydSGo9yAezrgLsq3jpm2p3zsb9bWo4tfx8dWz
+  contract_dispatcher.py: Qma29pRUzso927fUMrZoHSN46CTtv874o9N3p849PVi5bj
+  ledger_dispatcher.py: QmQEFDkXXvZgwjNNnGX8nzUS4mRaAYF7gRJFfxBqiSdvqv
 fingerprint_ignore_patterns: []
 protocols:
 - fetchai/contract_api:0.1.0
 - fetchai/ledger_api:0.1.0
 class_name: LedgerApiConnection
-config:
-  foo: bar
+config: {}
 excluded_protocols: []
 restricted_to_protocols:
 - fetchai/contract_api:0.1.0
