+++ packages/fetchai/protocols/contract_api/contract_api.proto
@@ -13,6 +13,10 @@ message ContractApiMessage{
         bytes signed_transaction = 1;
     }
 
+    message State{
+        bytes state = 1;
+    }
+
     message Terms{
         bytes terms = 1;
     }
@@ -45,8 +49,8 @@ message ContractApiMessage{
         string transaction_digest = 2;
     }
 
-    message Balance_Performative{
-        int32 balance = 1;
+    message State_Performative{
+        State state_data = 1;
     }
 
     message Raw_Transaction_Performative{
@@ -76,13 +80,13 @@ message ContractApiMessage{
     string dialogue_responder_reference = 3;
     int32 target = 4;
     oneof performative{
-        Balance_Performative balance = 5;
-        Error_Performative error = 6;
-        Get_Raw_Transaction_Performative get_raw_transaction = 7;
-        Get_State_Performative get_state = 8;
-        Get_Transaction_Receipt_Performative get_transaction_receipt = 9;
-        Raw_Transaction_Performative raw_transaction = 10;
-        Send_Signed_Transaction_Performative send_signed_transaction = 11;
+        Error_Performative error = 5;
+        Get_Raw_Transaction_Performative get_raw_transaction = 6;
+        Get_State_Performative get_state = 7;
+        Get_Transaction_Receipt_Performative get_transaction_receipt = 8;
+        Raw_Transaction_Performative raw_transaction = 9;
+        Send_Signed_Transaction_Performative send_signed_transaction = 10;
+        State_Performative state = 11;
         Transaction_Digest_Performative transaction_digest = 12;
         Transaction_Receipt_Performative transaction_receipt = 13;
     }
