+++ packages/fetchai/protocols/contract_api/custom_types.py
@@ -97,6 +97,42 @@ class SignedTransaction:
         raise NotImplementedError
 
 
+class State:
+    """This class represents an instance of State."""
+
+    def __init__(self):
+        """Initialise an instance of State."""
+        raise NotImplementedError
+
+    @staticmethod
+    def encode(state_protobuf_object, state_object: "State") -> None:
+        """
+        Encode an instance of this class into the protocol buffer object.
+
+        The protocol buffer object in the state_protobuf_object argument must be matched with the instance of this class in the 'state_object' argument.
+
+        :param state_protobuf_object: the protocol buffer object whose type corresponds with this class.
+        :param state_object: an instance of this class to be encoded in the protocol buffer object.
+        :return: None
+        """
+        raise NotImplementedError
+
+    @classmethod
+    def decode(cls, state_protobuf_object) -> "State":
+        """
+        Decode a protocol buffer object that corresponds with this class into an instance of this class.
+
+        A new instance of this class must be created that matches the protocol buffer object in the 'state_protobuf_object' argument.
+
+        :param state_protobuf_object: the protocol buffer object whose type corresponds with this class.
+        :return: A new instance of this class that matches the protocol buffer object in the 'state_protobuf_object' argument.
+        """
+        raise NotImplementedError
+
+    def __eq__(self, other):
+        raise NotImplementedError
+
+
 class Terms:
     """This class represents an instance of Terms."""
