+++ packages/hashes.csv
@@ -21,7 +21,7 @@ fetchai/agents/weather_station,QmUvYQUUka81UjcFRsWFxDvey77hQee6SdDYfJdufAvcq3
 fetchai/connections/gym,QmVdd7gaXKS5STrrifcKykcoh66BFHyb3THDRRMyTqRtLM
 fetchai/connections/http_client,Qmf3Mx9KtaQWiYsLNyBAKqAFntyi7qPSejFjKdUirQuGTa
 fetchai/connections/http_server,QmRQtynj999QJKbjMDRBdE5n55LaD9cRQ86i2zW1QqYykd
-fetchai/connections/ledger_api,QmZiCDb8AMXGa6LtxMVo5rdAuiJaDedCUEwty6NQha88Ly
+fetchai/connections/ledger_api,QmdKq3yGWmcg2LLbz9WznZy9DqPErWD9QzMAsgG8HGeaaq
 fetchai/connections/local,QmSNLjEDTchd3MaFhN7t8v8Gye52vKVWn1mBfHFJHbdiiS
 fetchai/connections/oef,QmSicSVkq6m1wSHRA9GKCAXFR2VGn9SAk2xNLXcin7v7a3
 fetchai/connections/p2p_client,QmTCX4D9JhtWufVqLAi8mJZH2eHj5WTaHzmJq5LKGEbUSF
@@ -30,13 +30,13 @@ fetchai/connections/p2p_libp2p_client,QmZAjbv1wUy8uQ2jLoDYbcr6PLFx8Yo6K5ykDiZJnM
 fetchai/connections/p2p_stub,QmZ9NCpe6Vs9TGEXLMNmcVQ197zAih17aDrBoqDC2B6TG6
 fetchai/connections/scaffold,QmQjE2UuxECHaehmoGgahRUqTGddfcN287UHZPe6jmyXof
 fetchai/connections/soef,QmSELgoVaUkxWDURsPTXS7zcAc2ETdS6SQKCD9fyGTNLAG
-fetchai/connections/stub,QmQtso71WsMsiF8F6UhNFUrc9x6EG85TMSGMxSVJVem7Sf
+fetchai/connections/stub,QmcJwjrjwFiAteo9QhV2AF62JF14sbczoiu3AG1EjLQUzK
 fetchai/connections/tcp,QmRuB5htAyYaWVQiSmYXqHL4MArzM9t14kRHKG4ZmkPePL
 fetchai/connections/webhook,Qma5WJq8CXajhYXaLz3cnaEzmqyPEouvvd6yuYosESXM9S
 fetchai/contracts/erc1155,QmPZqLiFxqU5ybchFSf3osCjjjUBQ4xThzAtdnUAZb8P3s
 fetchai/contracts/scaffold,QmemGGZ2znyWCqgr7jpS9aUYdVr1NH2NCnG9z2R8StxMKb
-fetchai/protocols/contract_api,QmeGktbv6rW6KdhQDRdjk2LZ3DRW5R2YfeKHAyng9T6rSN
-fetchai/protocols/default,QmaCMZ2udbarH46f7jydA6VvbVseqQRbmbqT3aJdivz1c2
+fetchai/protocols/contract_api,QmYLhxNNk6KVoNdvZ91UngJ1L2G2bFPkuXMzSqbSAuq4uQ
+fetchai/protocols/default,QmQWBpaM9bij3yjikdBSh7Su8ZVhBRqDBYtvuRFBgxNvfp
 fetchai/protocols/fipa,QmfEmaCogWuMzrDrvgy51tyYkaM5dJ7vywE2bVSzAEzZBe
 fetchai/protocols/gym,QmUNndNWuC9J4sxVWjV1dzsCCHFMdqZ6SoHWgN2rJcAonQ
 fetchai/protocols/http,Qmc4BkFqdjSRRN6HpSWzLNG5Lih32tG1x465XdP2uKwMdL
@@ -52,7 +52,7 @@ fetchai/skills/carpark_detection,QmTucyPTKnymjqfVdPfD9Zoasqs63LL9sC8jTuNq7BBEmC
 fetchai/skills/echo,Qmb5LwXSxHbibRp7X1pWXtrp3fzwAe4bChECYhLzkivuzk
 fetchai/skills/erc1155_client,QmTpn5aj5LAdUhY7vNQFAW1Z6oG93wgdivaht5zQAxDckw
 fetchai/skills/erc1155_deploy,QmVXpVjT8VnWBEaDD9HUWU7rzk1mVDzBCcapKguEifJEHn
-fetchai/skills/error,QmRE9DQr3jZN3afDeKbVTpBDfQjdWnyYJLApaz7YY1AMfc
+fetchai/skills/error,QmWHEXKeJT3vDDj6ZfuvE2ACvmgZKLdHmfDUNXpkmobuxW
 fetchai/skills/generic_buyer,QmVyyqxuBojiRXzUzgQrNckpJJAedJiia5uWCeeUHN1SzD
 fetchai/skills/generic_seller,Qmbe53BbgJmJauY7TtTpEMPofVEuuSm9MpqpJjLjFfTqUs
 fetchai/skills/gym,QmUYmjU7ejhxWTe45tcnDURqNZJCyNC1AQkzKHayhhQgv5
